refactoring userController name
