+++ api/routes/user.js
@@ -7,7 +7,7 @@ import {
   updateDetails,
   updatePassword,
   deleteUser,
-} from "../controllers/userController2.js";
+} from "../controllers/userController.js";
 import authorize from "../middleware/authorize.js";
 import {
   loginRules,
